Refactor components for RideConfirmedScreen
