+++ app/src/main/java/com/example/uberbookingexperience/ui/screens/rideConfirmed/components/RidePin.kt
@@ -1,0 +1,48 @@
+package com.example.uberbookingexperience.ui.screens.rideConfirmed.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import com.example.uberbookingexperience.ui.theme.spacing
+
+@Composable
+fun RidePin(title: String, pin: String) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(MaterialTheme.spacing.medium),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            text = title,
+            style = MaterialTheme.typography.titleMedium,
+            fontWeight = FontWeight.Normal
+        )
+        Row {
+            pin.forEach { pinCharacter ->
+                Text(
+                    modifier = Modifier
+                        .padding(end = MaterialTheme.spacing.extraSmall)
+                        .background(MaterialTheme.colorScheme.secondary)
+                        .padding(
+                            horizontal = MaterialTheme.spacing.small,
+                            vertical = MaterialTheme.spacing.extraSmall
+                        ),
+                    text = pinCharacter.toString(),
+                    color = MaterialTheme.colorScheme.surface,
+                    style = MaterialTheme.typography.bodyMedium,
+                    fontWeight = FontWeight.Bold
+                )
+            }
+        }
+    }
+}
